feat: 🎸 isEffectOf (weak dependencies)
Sometimes steps are only an effect of another; for example the
deployment of code coverage is only an effect of running the unit tests
- even though we add both steps to the pipeline, we really only expect
that the coverage step is executed if the test step is executed as well.
`isEffectOf` makes that possible, it allows you to specify a step as a
dependency without marking it as required (a weak dependency so to say).

Closes: #29