refactor(command step): API for parameter overrides is less opaque and just uses the raw value
Version 5.7.x and 5.8.x prefixed a given parameter override with $. This meant that the API for such
an override was quite awkward and assumed knowledge of buildkite-graph's implementation. Later
versions use the given value verbatim, assuming the env var format that Buildkite supports.

BREAKING CHANGE: `withParameterOverride` assumes a env var interpolation template to be passed (e.g.
starting with `$`)